Reimplement toggle
Toggle is now handled in TeleportManager.
